Added CRUD functionality and a list showing all the record transactions

+++ Projects/record_store_sales_tracker/src/main/java/info/hccis/recordstore/bo/RecordSaleBO.java
@@ -73,7 +73,7 @@ public class RecordSaleBO {
      */
     public ArtistSaleList processCost(ArtistSaleList recordSale) {
         double cost = calculateSaleAmount(recordSale.getUnitsSold());
-        recordSale.setSaleAmount(BigDecimal.valueOf(cost));
+        recordSale.setSaleAmount((float) cost); // convert double to float
         return recordSale;
     }
